requirements.txt
telethon==1.36.0
groq>=0.11.0
python-dotenv==1.0.1
loguru==0.7.2
